Output Results after running the Anthropic Study
Since importing in the entire anthropic misalignment study isn't needed(see relevant links in the report). We decided to upload the experiments folder we used which contains all of our prompts and outputs that we analyzed. the big change we made to the study was switching it to use chat-gpt 4 instead of claude as our tester.